+++ spec/Pim/Bundle/TransformBundle/Normalizer/Flat/AttributeOptionNormalizerSpec.php
@@ -46,6 +46,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
         $option->getCode()->willReturn('red');
         $option->getAttribute()->willReturn($attribute);
         $option->isDefault()->willReturn(true);
+        $option->getSortOrder()->willReturn(1);
         $attribute->getCode()->willReturn('color');
         $option->getOptionValues()->willReturn([
             'en_US' => $valueEn,
@@ -60,6 +61,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
             'attribute' => 'color',
             'code' => 'red',
             'default' => '1',
+            'sort_order' => 1,
             'label-en_US' => 'Red',
             'label-fr_FR' => 'Rouge',
             'label-de_DE' => '',
@@ -75,6 +77,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
         $option->getCode()->willReturn('red');
         $option->getAttribute()->willReturn($attribute);
         $option->isDefault()->willReturn(true);
+        $option->getSortOrder()->willReturn(1);
         $attribute->getCode()->willReturn('color');
         $option->getOptionValues()->willReturn([
             'en_US' => $valueEn,
@@ -89,6 +92,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
             'attribute' => 'color',
             'code' => 'red',
             'default' => '1',
+            'sort_order' => 1,
             'label-en_US' => 'Red',
             'label-de_DE' => '',
         ]);
@@ -104,6 +108,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
         $option->getCode()->willReturn('red');
         $option->getAttribute()->willReturn($attribute);
         $option->isDefault()->willReturn(true);
+        $option->getSortOrder()->willReturn(1);
         $attribute->getCode()->willReturn('color');
         $option->getOptionValues()->willReturn([
             'en_US' => $valueEn,
@@ -121,6 +126,7 @@ class AttributeOptionNormalizerSpec extends ObjectBehavior
             'attribute' => 'color',
             'code' => 'red',
             'default' => '1',
+            'sort_order' => 1,
             'label-en_US' => 'Red',
             'label-fr_FR' => 'Rouge',
             'label-de_DE' => '',